Add search page module and remove controller extension
The controller methods were available on every page in Alchemy and was listening to the query param. This approach is not really extendable, if you need multiple parameter to process the search result.
The main logic was moved into a SearchPage module which has nearly the same structure as before, but is now way more flexible, because it is only used in this particular page layout. It also allows the application to extend the search and the view without monkey patching the gem.
The only downside is, that it isn't possible anymore to add a search string the searchresults - element to test the search in the admin interface.